- Changed the database layout. The old database layout is incompatible. Please recreate an phTagr instance and reinsert the image! - Insert an Image classe to handle the data of the image. The file sync.php is now deprecated. - Started the ACL handling in the User class. This class checks all operation. E.g. is an user allowed to change a tag of an image. - Improve the session handling. E.g. the session stores the viewed images of an user.
git-svn-id: https://svn.phtagr.org/phtagr/trunk@70 c01cf435-4f0c-0410-981c-aea1cd8e29bc

+++ phtagr/phtagr/SectionUpload.php
@@ -344,24 +344,30 @@ function print_uploaded()
     {
       while (($file = readdir($dh)) !== false)
       {
-        if ($file !== '.' && $file !== '..' && !is_dir($file))
+        if ($file != '.' && 
+            $file != '..' && 
+            !is_dir($file) && 
+            file_exists($fullpath.$file))
         {
-          $sql = "SELECT id FROM ".$db->image."
-                  WHERE filename='". $fullpath . $file . "'";
-          $result = $db->query($sql);
-          $v = mysql_fetch_array ($result, MYSQL_ASSOC);
+          $image=new Image(-1);
+          if (!$image->init_by_filename($fullpath.$file))
+            continue;
 
           echo "    <tr>\n";
           echo "      <td align=\"center\" width=\"5\"><input type=\"checkbox\" name=\"".$file."\" /></td>\n";
           echo "      <td align=\"center\">\n";
-          echo print_mini($v['id']) . "</td>\n";
+          
+          $src=$image->create_mini();
+          echo "<img src=\"$src\">";
+          echo "</td>\n";
           echo "      <td align=\"center\">$file</td>\n";
           echo "      <td align=\"right\">"
-             . $this->get_readable_size ( filesize($fullpath . $file) ) 
+             . $this->get_readable_size ( $image->get_bytes() ) 
              . " Bytes</td>\n";
           echo "<td><div class=\"headerright\"><a href='./index.php?section=upload&delete_upload=&".urlencode($file)."=on'>Delete</a></div></td>\n";
           echo "    </tr>\n";
-          $file_size_sum += filesize($fullpath . $file) ;
+          $file_size_sum += $image->get_bytes() ;
+          unset($image);
         }
       }
       echo "    <tr>\n";
